requirements.txt
asgiref==3.5.0
Django==3.1.7
django-environ==0.8.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
